Dialogs (#30)
* Build gradle update

* Build gradle update

* Build gradle update

* Build gradle update

* Build gradle update

* Build gradle update

* Build gradle update

* Build gradle update

* Build gradle update

* Build gradle update

* Moving files

* Refactoring

* Build gradle update

* Build gradle update

* Build gradle update


* Build gradle update


* Build gradle update


* Build gradle update


* Build gradle update


* Dialogs Improvement


* Dialogs Improvement


* Dialogs Improvement


* Dialogs Improvement


* Dialogs Improvement


* Dialogs Improvement


* Dynamic layout bug fix


* State holders refactoring


* State holders refactoring


---------

+++ composeApp/src/iosMain/kotlin/sharedData/ScreenData.ios.kt
@@ -3,12 +3,12 @@ package sharedData
 import androidx.compose.runtime.Composable
 
 @Composable
-actual fun actualWidthFactor(): Int {
+actual fun actualWidthFactor(): Float {
     TODO("Not yet implemented")
 }
 
 @Composable
-actual fun actualHeightFactor(): Int {
+actual fun actualHeightFactor(): Float {
     TODO("Not yet implemented")
 }
 
@@ -21,5 +21,3 @@ actual fun getScreenWidth(): Int {
 actual fun getScreenHeight(): Int {
     TODO("Not yet implemented")
 }
-
-
